@@ -21,8 +21,6 @@ from urllib3 import HTTPResponse
 from urllib3.exceptions import MaxRetryError
 from urllib3.util import Retry
 
-from gretel_client._api.api_client import ApiClient as V2ApiClient
-from gretel_client._api.configuration import Configuration as V2Configuration
 from gretel_client.rest.api.projects_api import ProjectsApi
 from gretel_client.rest.api.users_api import UsersApi
 from gretel_client.rest.api_client import ApiClient
@@ -137,11 +135,9 @@ class PreviewFeatures(Enum):
 class GretelClientConfigurationError(Exception): ...
 
 
-T = TypeVar("T")
-ClientT = TypeVar("ClientT", bound=Union[ApiClient, V1ApiClient, V2ApiClient])
-ConfigT = TypeVar(
-    "ConfigT", bound=Union[Configuration, V1Configuration, V2Configuration]
-)
+T = TypeVar("T", bound=Type)
+ClientT = TypeVar("ClientT", bound=Union[ApiClient, V1ApiClient])
+ConfigT = TypeVar("ConfigT", bound=Union[Configuration, V1Configuration])
 
 
 class RunnerMode(str, Enum):
@@ -278,13 +274,6 @@ class ClientConfig(ABC):
     def endpoint(self) -> str: ...
 
     @property
-    def console_endpoint(self) -> str:
-        if self.stage == "dev":
-            return "https://console-eng.gretel.ai"
-        else:
-            return "https://console.gretel.ai"
-
-    @property
     @abstractmethod
     def artifact_endpoint(self) -> str: ...
 
@@ -317,7 +306,6 @@ class ClientConfig(ABC):
         if (
             "https://api-dev.gretel" in self.endpoint
             or ".dev.gretel.cloud" in self.endpoint
-            or ".dev.gretel.ai" in self.endpoint
         ):
             return "dev"
         return "prod"
@@ -460,7 +448,6 @@ class DefaultClientConfig(ClientConfig):
     def __init__(
         self,
         endpoint: Optional[str] = None,
-        console_endpoint: Optional[str] = None,
         artifact_endpoint: Optional[str] = None,
         api_key: Optional[str] = None,
         default_project_name: Optional[str] = None,
@@ -620,10 +607,6 @@ class DelegatingClientConfig(ClientConfig):
     @property
     def endpoint(self) -> str:
         return self._delegate.endpoint
-
-    @property
-    def console_endpoint(self) -> str:
-        return self._delegate.console_endpoint
 
     @property
     def artifact_endpoint(self) -> str:
