@@ -85,12 +85,8 @@ class WorkflowTaskError(Exception):
 @dataclass
 class Message:
 
-    step: Optional[str]
-    """
-    The name of the step.
-    
-    If the message is not associated with a step, this value will be `None`.
-    """
+    step: str
+    """The name of the step"""
 
     stream: str
     """
@@ -111,9 +107,6 @@ class Message:
     @classmethod
     def from_dict(cls, message: dict, raise_on_error: bool = False) -> Message:
         message["ts"] = datetime.fromisoformat(message["ts"])
-
-        if "step" not in message:
-            message["step"] = None
         deserialized_message = cls(**message)
 
         if raise_on_error:
