@@ -1,6 +1,6 @@
 from typing import Any, Optional, Union
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 from gretel_client.navigator.data_designer.data_column import GeneratedDataColumn
 from gretel_client.navigator.tasks.types import (
@@ -25,7 +25,7 @@ class PostProcessEvaluator(PostProcessor):
 
 class AIDDConfig(BaseModel):
     model_suite: ModelSuite
-    categorical_seed_columns: list[SeedCategory] = Field(..., min_length=1)
+    categorical_seed_columns: list[SeedCategory]
     generated_data_columns: list[GeneratedDataColumn]
     post_processors: Optional[
         list[Union[PostProcessValidator, PostProcessEvaluator]]
